requirements.txt
streamlit>=1.39.0
python-dotenv>=1.0.1
groq>=0.9.0
duckduckgo-search>=6.2.6
wikipedia>=1.4.0
arxiv>=2.1.0
youtube-search-python>=1.6.6
requests>=2.31.0
openai>=1.47.0
google-generativeai>=0.7.2
ddgs